SysUI:Custom Recents Style -TV header [1/2]
Task View Header Customizations
Earlier we used to have Icons On the Task View Header to be either dark or light depending upon the Theme's action bar
Now lets override this to define custom colors to these drawables . we also provide a switch to maintain integrity with themes. this switch is linked to custom recent Styles.
Also We Tint Entire App Icon to make it look more material. if app icon needs to be tinted with viewing its elements, set transperancy to the icon and its default color to be 0x00ffffff

This Custom Recents Styles is a continuation to my 2 commits

https://github.com/ResurrectionRemix/android_frameworks_base/commit/7930f572cb5840b363f6292e4946471d69a6289c

and

https://github.com/ResurrectionRemix/android_frameworks_base/commit/980d0dc2152a285247e915f5dc6adbf859b708c4

--Also Update FW drawable

Credits :
Google for their MW implementation
Slim Roms for Floating Window Implementation and drawable
OmniRom for Dark and Light MW Adjustments

Signed-off-by: Varun Date <date.varun123@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskViewHeader.java